Move uncaught exception handler outside AppCenter class
To reduce size of file.
